add parent class classifier.py: all models should inherit from it
add main methods : train, predict, error, error_pred (error between prediction and ground truth on a batch), verbose (description of a model). Main focus was the cross validation method: in order to make it usable by all models, the function shouldnt take the number of hyperparameters as argument. Recursive framework was used to bypass the problem